CMake test target if php is found, MDEL on big data test, better test handling with command line arguments

+++ tests/run_tests.php
@@ -5,35 +5,58 @@ if( !class_exists('Gibson') ) die( "You will need the Gibson PHP extension to ru
 
 chdir( dirname(__FILE__) );
 
-system("cd .. && cmake . && make && ./gibson -c tests/config.test");
+$shortopts = "C:S:"; 
+$longopts  = array(
+	"custom-server:",
+	"single-test:" 
+);
+$options = getopt($shortopts, $longopts);
 
-define( 'GIBSON_SOCKET', '../gibson.sock' );
+if( isset( $options['C'] ) || isset( $options['custom-server'] ) ){
+	$bCustomServer = TRUE;
+	define( 'GIBSON_SOCKET', isset( $options['C'] ) ? $options['C'] : $options['custom-server'] );
+}
+else {
+	$bCustomServer = FALSE;
+	define( 'GIBSON_SOCKET', '../gibson.sock' );
+}
+
+if( isset( $options['S'] ) || isset( $options['single-test'] ) ){
+	$tests = array( ( isset( $options['S'] ) ? $options['S'] : $options['single-test'] ).'.test.php' );
+}
+else {
+	$tests = glob("*.test.php" );
+}
+
+if( $bCustomServer == FALSE )
+	system("cd .. && cmake . && make && ./gibson -c tests/config.test");
 
 $passed = 0;
 $failed = 0;
 $start  = microtime(TRUE);
 
-foreach( glob("*.test.php" ) as $testfile ){
+foreach( $tests as $testfile ){
 	$name = strtoupper( str_replace( '.test.php', '', $testfile ) );
 	
 	printf( "Running %10s test ...", $name );
   
-  $t_start = microtime(TRUE);
+ 	$t_start = microtime(TRUE);
 	$output  = trim( shell_exec("php $testfile") );
-  $t_end   = microtime(TRUE);
+  	$t_end   = microtime(TRUE);
 	if( $output ){
 		printf( " FAILED: %s\n", $output );
-    ++$failed;  
-  }
+   		++$failed;  
+  	}
 	else { 
 		printf( " PASSED ( %s )\n", format_timespan( $t_end - $t_start ) ); 
-    ++$passed;
-  }
+    	++$passed;
+  	}
 }
 
 $end = microtime(TRUE);
 
-system("killall gibson");
+if( $bCustomServer == FALSE )
+	system("killall gibson");
 
 printf( "\n%d passed, %d failed ( %s )\n", $passed, $failed, format_timespan( $end - $start ) );
